// src/interfaces/TableRowInterface.tsx
interface TableRowInterface {
  id: string;
  companySigDate?: string;
  companySignatureName: string;
  documentName: string;
  documentStatus: string;
  documentType: string;
  employeeNumber: string;
  employeeSigDate?: string;
  employeeSignatureName: string;
}

export default TableRowInterface;